modules/contact/requirements/requirements.txt
#!chuck_extends requirements/requirements.txt

#!chuck_appends REQUIREMENTS
django-honeypot==0.3.0
#!end
